requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
sentry_sdk==2.14.0














#pipreqs /home/trinh/PycharmProjects/User_Management/user_management --force

qrcode~=8.0
pillow~=11.0.0
djangorestframework-simplejwt~=5.3.1
sentry-sdk~=2.14.0
django-cors-headers~=4.6.0
uuid~=1.30
celery~=5.4.0
redis~=5.2.1